requirements.txt
python-dotenv
pandas
numpy
setuptools
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn

# -e .
